Move tests to modules
