+++ src/__tests__/fixtures/components/_no-parent.scss
