+++ src/__tests__/fixtures/main2.scss
