+++ src/io/__tests__/writeFiles.spec.ts
@@ -1,6 +1,7 @@
 /* global describe:true */
-import { IFile } from "../../src/io/file";
-import { writeFiles as write } from "../../src/io/writeFiles";
+import { IFile } from "../file";
+import { writeFiles as write } from "../writeFiles";
+import * as Bluebird from "bluebird";
 import { assert as t } from "chai";
 import * as fs from "fs";
 import "mocha";
