+++ src/plugins/node-sass/__tests__/compile.spec.ts
@@ -1,6 +1,6 @@
-import { IFile } from "../../../src/io/file";
-import { compile } from "../../../src/plugins/node-sass/compile";
-import { getFixture } from "../../helpers";
+import { getFixture } from "../../../__tests__/helpers";
+import { IFile } from "../../../io/file";
+import { compile } from "../compile";
 import { assert as t } from "chai";
 import * as fs from "fs";
 import "mocha";
