+++ src/plugins/postcss/__tests__/index.spec.ts
@@ -1,4 +1,4 @@
-import * as postcss from "../../../src/plugins/postcss/index";
+import * as postcss from "../index";
 import { assert as t } from "chai";
 import "mocha";
